functions_llama_pipeline/requirements.txt
# functions_llama_pipeline — LlamaIndex Cloud RAG Pipeline (beta)
# Separate Firebase Functions codebase; does NOT conflict with functions-python.

# ── Firebase runtime ─────────────────────────────────────────────────────────
firebase-functions>=0.4.2,<1.0.0
firebase-admin>=6.5.0,<7.0.0

# ── Google Cloud infrastructure ──────────────────────────────────────────────
google-cloud-storage<3.0.0
google-cloud-firestore<3.0.0

# ── LlamaIndex Cloud + LlamaParse ────────────────────────────────────────────
llama-cloud>=1.0.0,<2.0.0
llama-parse>=0.6.0,<1.0.0
llama-index-core>=0.12.0,<1.0.0
llama-index-embeddings-openai>=0.3.0,<1.0.0
llama-index-llms-openai>=0.3.0,<1.0.0
llama-index-postprocessor-cohere-rerank>=0.3.0,<1.0.0

# ── OpenAI (embedding + LLM fallback) ───────────────────────────────────────
openai>=1.23.0,<2.0.0

# ── Async support ────────────────────────────────────────────────────────────
nest-asyncio>=1.6.0,<2.0.0
